ByteConverter added, Sockets creation changed

+++ src/include/M4DCommon.h
@@ -14,7 +14,7 @@
  * typedef of basic data types for simplier porting
  */
 // signed
-typedef char int8;
+typedef signed char int8;
 typedef short int16;
 typedef int int32;
 typedef long int64;
@@ -22,7 +22,7 @@ typedef long int64;
 typedef unsigned char uint8;
 typedef unsigned short uint16;
 typedef unsigned int uint32;
-typedef unsigned long uint64;
+typedef unsigned long long uint64;
 // floats
 typedef float float32;
 typedef double float64;
